chore: type StyledSelectFilter
